backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai==1.3.7
langchain==0.0.340
langchain-openai==0.0.2
python-dotenv==1.0.0
httpx==0.25.2
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
